requirements.txt
fastapi
uvicorn
pinecone-client
pandas
python-dotenv
transformers
torch
together
tqdm
requests
